+++ examples/high_level/complete_graph_tdvp.jl
@@ -6,7 +6,7 @@ limits = Limits(
 )
 
 output(n) =
-    "data" => [X, Y, Z, Y(1)Y(2), Z(1)Y(2)Y(3), EE(n ÷ 2, 4), Purity, Trace]
+    "data" => [X, Y, Z, Y(1)Y(2), Z(1)Y(2)Y(3), EntanglementEntropy(n ÷ 2, 4), Purity, Trace]
 
 sim_data(n) = SimData(
     name = "my_simulation_with_$(n)_qubits",