+++ src/Phases.jl
@@ -164,7 +164,7 @@ function run_phase(sim::Simulation, phase::SteadyState)
         return sim
     end
     e, sim = steady_state(phase.lindbladian, sim;
-        phase.nsweeps, first_sweep = done + 1, phase.limits, phase.mpo_limits, alg = phase.mpo_algo,
+        phase.nsweeps, first_sweep = done + 1, phase.limits, phase.mpo_limits, phase.mpo_algo,
         observer! = DmrgObserver(sim, phase.measures, phase.measures_period, phase.tolerance, done))
     log_msg(sim, "Done, dmrg final value is $e (0 for steady state)")
     return sim